MINOR: changes to lesson 12
These changes provide the following:

 - explain that there is actually an `Extension` class
 - added explicity `=1` to flush

I would also look at moving the traits explanation to the bottom of that section. Explain why extensions are cool before going on the defensive?

The 'owner gotcha' could also be rewritten to explain how everything that is on the master class should be accessed via "$this->owner->Blah" rather than it seeming like an addendum to the paragraph? Its more important to know how to use the system than being told when you'll screw up (and I still don't love the error screen).